Hot Stone Massage.
Those who understand you also understand the connections they want. Focusing on communication is very important, and hot stone massages offer to dedicate time to each other as you relax. Thank you for the opportunity to soak up the positive waves, ponder your feelings and experiences, and enjoy the partnership. Hot Stone Massage is a soft, flat heated stone placed on a specific part of the body. The stone is usually composed of basalt, which retains heat in a basalt type of volcanic rock.